refactor: Change 404 feature to _error page
